Fix wrong initial memory validation size

+++ wasm_doctor.h
@@ -7,8 +7,8 @@
 typedef uint32_t wasmptr_t;
 
 void move_shadow_stack_pointer(wasmptr_t address);
-void shadow_store(wasmptr_t address, uint8_t size);
-bool shadow_load(wasmptr_t address, uint8_t size);
+void shadow_store(wasmptr_t address, uint32_t size);
+bool shadow_load(wasmptr_t address, uint32_t size);
 void doctor_init(uint32_t num_of_pages);
 
 #endif /* WASM_DOCTOR */
